+++ FalcLib/F4VU.cs
@@ -5,7 +5,7 @@ using log4net;
 using FalconNet.F4Common;
 using VU_BYTE = System.Byte;
 using VU_TIME = System.UInt64;
-using VU_ID_NUMBER = System.UInt64;
+using VU_ID_NUMBER = System.UInt32;
 using VU_KEY = System.UInt64;
 using VU_BOOL = System.Boolean;
 using BIG_SCALAR = System.Single;
